requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1743190139881/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
archspec @ file:///croot/archspec_1709217642129/work
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
conda @ file:///croot/conda_1743715288469/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1744662537425/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1744823864153/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1740577825284/work
cycler==0.12.1
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
filelock==3.19.1
fonttools==4.58.5
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
fsspec==2025.7.0
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
imageio==2.37.0
Jinja2==3.1.6
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
kiwisolver==1.4.8
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1738943416351/work
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging @ file:///croot/packaging_1734472117206/work
pillow==11.3.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
requests @ file:///croot/requests_1730999120400/work
rich @ file:///croot/rich_1732638981168/work
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
scikit-image==0.25.2
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.6.11
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm @ file:///croot/tqdm_1738943501192/work
triton==3.4.0
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wheel==0.45.1
zstandard @ file:///croot/zstandard_1731356346222/work
